+++ CHANGELOG.md
@@ -34,6 +34,8 @@ version 0.2.0, thus covering only a fraction of the project’s history.
     - M3 screws instead of M4 in `config/base.yaml`.
 
 ### New
+- Added support Kailh’s PG1511 series switches and similar MX clones without
+  lateral recesses in the lower body.
 - Added a nominal clearance parameter to make it easier to design for multiple
   different types of switches and keycaps.
 - Added support for custom key mounting plate size.